+++ src/constant.js
@@ -1,5 +1,5 @@
 const port = process.env.REACT_APP_API_PORT || 1337;
-export const API = `http://192.168.31.36:${port}/api`;
+export const API = `http://localhost:${port}/api`;
 export const AUTH_TOKEN = "authToken";
 export const BEARER = "Bearer";
 export const ACTIVITY_CATEGORIES = {
